timberland: change the libnvme submodule and update README
The Timberland-sig project is using a copy
of the nvme-cli and libnvme repositories until release.

This change sets the libnvme submodule to timberland-sig
and updates the README.md file with work flow instructions.

The default/feature branch is now name timberlandmaster.
This applies to both nvme-cli and libnvme.

Add instructions how to override system libnvme

Update libnvme.wrap to use timberlandmaster

Note: when merging upstream, remove this sha.

Signed-off-by: John Meneghini <jmeneghi@redhat.com>